size_t

+++ trunk/fflas-ffpack/fflas/fflas_igemm/igemm.h
@@ -81,11 +81,11 @@ namespace FFLAS { namespace Protected {
 namespace FFLAS { /*  igemm */
 
 	inline void igemm_(const enum FFLAS_ORDER Order, const enum FFLAS_TRANSPOSE TransA, const enum FFLAS_TRANSPOSE TransB
-			   , const int M, const int N, const int K
+			   , const size_t M, const size_t N, const size_t K
 			   , const int64_t alpha
-			   , const int64_t *A, const int lda, const int64_t *B, const int ldb
+			   , const int64_t *A, const size_t lda, const int64_t *B, const size_t ldb
 			   , const int64_t beta
-			   , int64_t *C, const int ldc);
+			   , int64_t *C, const size_t ldc);
 
 
 } // FFLAS
